PFEH2025 - ajustements des tests problematique
